Split Repository into FeedService/TimelineService interfaces
Breaks the monolithic Repository interface in internal/seymour into
per-domain FeedService and TimelineService interfaces (feed.go,
timeline.go), and updates all dependents (api, worker) accordingly.
