+++ internal/seymour/seymour.go
@@ -1,111 +1,9 @@
 package seymour
 
 import (
-	"context"
 	"database/sql/driver"
 	"fmt"
 	"time"
-)
-
-// Repository provides a unified interface for all data operations.
-// It combines the functionality of FeedService and TimelineService
-// into a single interface to simplify dependency injection and testing.
-type Repository interface {
-	// Feed operations
-	Feed(ctx context.Context, id string) (Feed, error)
-	Feeds(ctx context.Context, ids []string) ([]Feed, error)
-	FeedByURL(ctx context.Context, url string) (Feed, error)
-	InsertFeed(ctx context.Context, url string) (Feed, error)
-	DeleteFeed(ctx context.Context, id string) error
-	CountAllFeeds(ctx context.Context) (int, error)
-	FeedIDs(ctx context.Context, offset, pageSize int) ([]string, error)
-	Entry(ctx context.Context, id string) (FeedEntry, error)
-	Entries(ctx context.Context, ids []string) ([]FeedEntry, error)
-	InsertEntries(ctx context.Context, entries []FeedEntry) error
-	UpdateFeed(ctx context.Context, id string, args UpdateFeedArgs) error
-
-	// Timeline operations
-	CreateSubscription(ctx context.Context, feedID string) error
-	AllSubscriptions(ctx context.Context) ([]Subscription, error)
-	MissingEntries(ctx context.Context) ([]MissingEntry, error)
-	EntriesNeedingJudgement(ctx context.Context, limit uint) ([]TimelineEntry, error)
-	InsertEntry(ctx context.Context, entry TimelineEntry) error
-	UpdateTimelineEntry(ctx context.Context, id string, status TimelineEntryStatus) error
-	TimelineEntries(ctx context.Context, args TimelineEntriesArgs) ([]TimelineEntry, error)
-	CountTimelineEntries(ctx context.Context, args TimelineEntriesArgs) (int, error)
-}
-
-// Feed represents an RSS feed's details.
-type Feed struct {
-	ID           string  `db:"id"`
-	Title        *string `db:"title"`
-	URL          string  `db:"url"`
-	Description  *string `db:"description"`
-	LastSyncedAt *DBTime `db:"last_synced_at"`
-	CreatedAt    DBTime  `db:"created_at"`
-	UpdatedAt    DBTime  `db:"updated_at"`
-}
-
-// FeedEntry represents a unique entry in an RSS feed.
-type FeedEntry struct {
-	ID          string `db:"id"`
-	FeedID      string `db:"feed_id"`
-	GUID        string `db:"guid"`
-	Title       string `db:"title"`
-	Description string `db:"description"`
-	CreatedAt   DBTime `db:"created_at"`
-	PublishTime DBTime `db:"publish_time"`
-	Link        string `db:"link"`
-}
-
-// UpdateFeedArgs holds the optional fields for updating a feed.
-type UpdateFeedArgs struct {
-	Title       string
-	Description string
-	LastSynced  DBTime
-}
-
-// Subscription represents a subscription to a feed.
-type Subscription struct {
-	ID        string `db:"id"`
-	FeedID    string `db:"feed_id"`
-	CreatedAt DBTime `db:"created_at"`
-}
-
-// TimelineEntry represents an entry in the timeline.
-type TimelineEntry struct {
-	ID          string `db:"id"`
-	FeedEntryID string `db:"feed_entry_id"`
-	CreatedAt   DBTime `db:"created_at"`
-	FeedID      string `db:"feed_id"`
-
-	// For curation: whether the entry has been approved by the judge
-	Status TimelineEntryStatus `db:"status"`
-}
-
-// MissingEntry is an instance where a feed entry should have been added to the timeline.
-type MissingEntry struct {
-	FeedEntryID string `db:"feed_entry_id"`
-	FeedID      string `db:"feed_id"`
-}
-
-// TimelineEntriesArgs holds arguments for filtering timeline entries.
-type TimelineEntriesArgs struct {
-	Status TimelineEntryStatus // To optionally filter by status
-	FeedID string              // To optionally filter by feed
-	Limit  uint64              // To optionally limit the number of entries returned
-
-	// Pagination fields
-	Offset uint64 // Offset for pagination
-}
-
-// TimelineEntryStatus represents the status of a timeline entry.
-type TimelineEntryStatus string
-
-const (
-	TimelineEntryStatusRequiresJudgement TimelineEntryStatus = "requires_judgement"
-	TimelineEntryStatusApproved          TimelineEntryStatus = "approved"
-	TimelineEntryStatusRejected          TimelineEntryStatus = "rejected"
 )
 
 // DBTime is a sqlite-acceptable implementation of a time that can be marshaled in and out of
